requirement.txt
annotated-types==0.6.0
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-ninja==1.1.0
dnspython==2.6.1
email_validator==2.1.1
idna==3.6
pillow==10.3.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
